FixingCalculatorOutputText

+++ Sem-2/CS1804T - Data Structures Lab/Lab_1_(24_1_24)/calculator.h
@@ -10,6 +10,6 @@ int Sub(int *num1 , int *num2){
 int Mul(int *num1 , int *num2){
     return *num1 * *num2;
 }
-int Div(int *num1 , int *num2){
-    return (*num1 / *num2);
+float Div(int *num1 , int *num2){
+    return (float)*num1 / *num2;
 }
